Telephone mount modification for "Wicked Chili" windshield mount.
The device mount was too far away, when mounted on the windshield with the standard suction cup.
Plus, the phone tends to run hot when placed there (sun).
I was looking for a solution to place the phone in front of a ventilation outlet. 
So I modified what I already had.
The lasered parts are slotted to fit into each other (3mm plywood). The "auflageflaeche" part is taped to a horizontal space beneath the car's digital clock display. The spacers ("abstandshalter") were lasered from acrylic. Plywood didn't do the job. These connect the new parts to the "old" ones.